+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 0.1.9 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fixed modal timeout
+  [eikichi18]
 
 
 0.1.8 (2018-07-19)